requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
certifi==2024.12.14
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
charset-normalizer==3.4.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==2.1.0
    # via -r requirements/test.in
coverage[toml]==7.6.10
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-waffle
    #   djangorestframework
    #   edx-django-utils
    #   openedx-filters
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
djangorestframework==3.15.2
    # via -r requirements/base.txt
edx-django-utils==7.1.0
    # via -r requirements/base.txt
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
newrelic==10.4.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
openedx-atlas==0.6.2
    # via -r requirements/base.txt
openedx-filters @ git+https://github.com/awais786/openedx-filters.git@course-templates-filters
    # via -r requirements/base.txt
packaging==24.2
    # via pytest
path==17.1.0
    # via -r requirements/base.txt
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
psutil==6.1.1
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
requests==2.32.3
    # via -r requirements/base.txt
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
text-unidecode==1.3
    # via python-slugify
urllib3==2.2.3
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   requests
django-user-tasks==3.2.0
